+++ src/test/pages/WorkbookDownload.test.tsx
@@ -1,58 +1,26 @@
-import { fireEvent, screen, waitFor } from "@testing-library/react";
-import { TestUserRole, checkPageTitle, renderTestEnvironment } from "../utils";
+import { fireEvent, waitFor } from "@testing-library/react";
+import {
+  checkPageTitle,
+  extraDownloadEndpoints,
+  getDownloadButtons,
+  renderTestEnvironment,
+} from "../utils";
 import { WorkbookDownload } from "../../app/components/pages/WorkbookDownload";
-import { API_PATH } from "../../app/services";
-import { rest } from "msw";
+import { DefaultBodyType, MockedRequest, RestHandler } from "msw";
 import * as actions from "../../app/state/actions";
 import * as download from "../../app/components/handlers/downloadWorkbook";
 
 const errorMessageSpy = jest.spyOn(actions, "showAxiosErrorToastIfNeeded");
 const downloadSpy = jest.spyOn(download, "downloadWorkbook");
 
-const getButtons = () => {
-  const gcseButton = () =>
-    screen.getByRole("button", {
-      name: /download gcse workbook/i,
-    });
-  const aqaButton = () =>
-    screen.getByRole("button", {
-      name: /download aqa workbook/i,
-    });
-  const ocrButton = () =>
-    screen.getByRole("button", {
-      name: /download ocr workbook/i,
-    });
-  return { gcseButton, aqaButton, ocrButton };
-};
-
-const renderWorkbookDownload = (role: TestUserRole) => {
+const renderWorkbookDownload = (
+  endpoints: RestHandler<MockedRequest<DefaultBodyType>>[]
+) => {
   renderTestEnvironment({
-    role: role,
+    role: "TEACHER",
     PageComponent: WorkbookDownload,
     initialRouteEntries: ["/workbook-download"],
-    extraEndpoints: [
-      rest.get(
-        API_PATH +
-          `/documents/content/books/gcse_book_23/isaac_cs_gcse_book_2023.pdf`,
-        (req, res, ctx) => {
-          return res(ctx.status(200), ctx.json({ data: "this is a book" }));
-        }
-      ),
-      rest.get(
-        API_PATH +
-          `/documents/content/books/workbook_20_aqa/isaac_cs_aqa_book_2022.pdf`,
-        (req, res, ctx) => {
-          return res(ctx.status(200), ctx.json({ data: "this is a book" }));
-        }
-      ),
-      rest.get(
-        API_PATH +
-          `/documents/content/books/workbook_20_ocr/isaac_cs_ocr_book_2022.pdf`,
-        (req, res, ctx) => {
-          return res(ctx.status(200), ctx.json({ data: "this is a book" }));
-        }
-      ),
-    ],
+    extraEndpoints: endpoints,
   });
 };
 
@@ -60,9 +28,9 @@ const buttons = ["GCSE", "OCR", "AQA"];
 
 describe("WorkbookDownload Component", () => {
   it("renders the buttons", () => {
-    renderWorkbookDownload("TEACHER");
+    renderWorkbookDownload(extraDownloadEndpoints.working);
     checkPageTitle("Isaac CS Workbook PDF Downloader");
-    const { gcseButton, aqaButton, ocrButton } = getButtons();
+    const { gcseButton, aqaButton, ocrButton } = getDownloadButtons();
     [gcseButton(), aqaButton(), ocrButton()].forEach((button) => {
       expect(button).toBeInTheDocument();
     });
@@ -71,8 +39,8 @@ describe("WorkbookDownload Component", () => {
   it.each(buttons)(
     "clicking on the button requests a download for the %s workbook",
     async (button) => {
-      renderWorkbookDownload("TEACHER");
-      const { gcseButton, aqaButton, ocrButton } = getButtons();
+      renderWorkbookDownload(extraDownloadEndpoints.working);
+      const { gcseButton, aqaButton, ocrButton } = getDownloadButtons();
       const buttonElement = {
         GCSE: gcseButton(),
         AQA: aqaButton(),
@@ -86,30 +54,8 @@ describe("WorkbookDownload Component", () => {
   );
 
   it("displays an error popup if handleDownload encounters an error", async () => {
-    renderTestEnvironment({
-      role: "TEACHER",
-      PageComponent: WorkbookDownload,
-      initialRouteEntries: ["/workbook-download"],
-      extraEndpoints: [
-        rest.get(
-          API_PATH +
-            `/documents/content/books/gcse_book_23/isaac_cs_gcse_book_2023.pdf`,
-          (req, res, ctx) => {
-            return res(
-              ctx.status(404),
-              ctx.json({
-                responseCode: 404,
-                responseCodeType: "Not Found",
-                errorMessage:
-                  "Unable to locate the file: content/books/gcse_book_23/isaac_cs_gcse_book_2023.pdf.",
-                bypassGenericSiteErrorPage: false,
-              })
-            );
-          }
-        ),
-      ],
-    });
-    const { gcseButton } = getButtons();
+    renderWorkbookDownload(extraDownloadEndpoints.broken);
+    const { gcseButton } = getDownloadButtons();
     fireEvent.click(gcseButton());
     await waitFor(() => {
       expect(errorMessageSpy).toHaveBeenCalled();
